Updated Manage Sections, Teacher Home, Student Dashboard, and Teacher Navbar
